PSXMLProtocol fixed xmlpp Document crash
As it turns out, xmlpp parser maintains a Document internally.
When the parser parses a new xml stream, it creates a new
instance of the internal Document, and deletes the previous one.
Therefore, the xmlpp Documents don't exist are gone by the time
the XML is returned from the XMLPSPrococol object.

To address this issue, the XMLPSProtocol object creates a new
object for each incoming XML message.  The object is allocated
on the heap and resource managed by boost's shared_ptr.  Hence
the API changed to include boost shared pointers.